chore: move node stuff + package.json changes
