+++ packages/utils/build.config.ts
@@ -25,8 +25,9 @@ export default defineBuildConfig({
 		{ input: 'src/loader/fs', name: 'loader/fs' },
 		{ input: 'src/loader/install-pkg', name: 'loader/install-pkg' },
 		{ input: 'src/loader/loader', name: 'loader/loader' },
-		{ input: 'src/loader/loaders', name: 'loader/loaders' },
 		{ input: 'src/loader/modern', name: 'loader/modern' },
+		{ input: 'src/loader/node-loader', name: 'loader/node-loader' },
+		{ input: 'src/loader/node-loaders', name: 'loader/node-loaders' },
 		{ input: 'src/loader/types', name: 'loader/types' },
 		{ input: 'src/loader/utils', name: 'loader/utils' },
 		{ input: 'src/loader/warn', name: 'loader/warn' },
