+++ packages/utils/jest.esm.config.ts
@@ -1,0 +1,7 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+	verbose: true,
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	testMatch: ['**/tests/*-test.ts'],
+};
